Move the link highlight code to the -typing,
rename findLinkRange to findAttributeRange,
so it could be shared between attributes.
